Modularizing the code into features.
